+++ vcxproj2cmake/Converter.cs
@@ -18,7 +18,7 @@ public class Converter
         List<FileInfo>? projectFiles = null,
         FileInfo? solutionFile = null,
         int? qtVersion = null,
-        Compiler compiler = Compiler.Msvc,
+        bool portable = false,
         bool includeHeaders = false,
         bool enableStandaloneProjectBuilds = false,
         IndentStyle indentStyle = IndentStyle.Spaces,
@@ -88,7 +88,7 @@ public class Converter
 
             try
             {
-                cmakeProjects.Add(new CMakeProject(project, new(qtVersion, compiler), includeHeaders, conanPackageInfoRepository, logger));
+                cmakeProjects.Add(new CMakeProject(project, new(qtVersion, portable), includeHeaders, conanPackageInfoRepository, logger));
             }
             catch (Exception ex) when (continueOnError)
             {